+++ src/test/java/org/morphix/reflection/MethodsGetGenericReturnClassTest.java
@@ -12,24 +12,21 @@
  */
 package org.morphix.reflection;
 
-import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.lang.reflect.Method;
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Type;
 import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
 /**
- * Test class for {@link Methods#getGenericReturnType(Method, int)}.
+ * Test class for {@link Methods#getGenericReturnClass(Method, int)}.
  *
  * @author Radu Sebastian LAZIN
  */
-class MethodsGetGenericReturnTypeTest {
+class MethodsGetGenericReturnClassTest {
 
 	public static class A {
 
@@ -48,30 +45,23 @@ class MethodsGetGenericReturnTypeTest {
 	}
 
 	@Test
-	void shouldReturnCorrectType() throws Exception {
+	void shouldReturnCorrectClass() throws Exception {
 		Method method = A.class.getMethod("getList1");
-		Type cls = Methods.getGenericReturnType(method, 0);
+		Class<?> cls = Methods.getGenericReturnClass(method, 0);
 
 		assertThat(cls, equalTo(String.class));
 	}
 
 	@Test
-	void shouldReturnParameterizedTypeFromParameterizedClass() throws Exception {
+	void shouldFailToCastFromParameterizedClass() throws Exception {
 		Method method = A.class.getMethod("getList2");
-		Type type = Methods.getGenericReturnType(method, 0);
-
-		assertThat(type, instanceOf(ParameterizedType.class));
+		assertThrows(ReflectionException.class, () -> Methods.getGenericReturnClass(method, 0));
 	}
 
 	@Test
 	void shouldFailForRawTypes() throws Exception {
 		Method method = A.class.getMethod("getList3");
-		ReflectionException e = assertThrows(ReflectionException.class, () -> Methods.getGenericReturnType(method, 0));
-
-		assertThat(e.getMessage(),
-				equalTo(method.getGenericReturnType().getTypeName() + " is a raw return type for method "
-						+ method.getDeclaringClass().getCanonicalName()
-						+ "." + method.getName()));
+		assertThrows(ReflectionException.class, () -> Methods.getGenericReturnClass(method, 0));
 	}
 
 }